backend/requirements.txt
flask
flask-cors
werkzeug
numpy
scikit-learn
scipy
python-docx
PyPDF2
openai
python-dotenv 